Extract 3D snapshot matcher into its own module
