+++ tests/fixtures/get-test-fixture.ts
@@ -3,6 +3,7 @@ import { RootCircuit } from "lib/RootCircuit"
 import { logSoup } from "@tscircuit/log-soup"
 import "lib/register-catalogue"
 import "./extend-expect-circuit-snapshot"
+import "./extend-expect-3d-matcher"
 import { preventExternalApiRequests } from "./prevent-external-api-requests"
 import { getTestStaticAssetsServer } from "./get-test-static-assets-server"
 import type { PlatformConfig } from "@tscircuit/props"
